requirements.txt
google-genai
streamlit
dotenv